+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Fixed
 
 - Error importing LibraryThing JSON when date wasn't specified
+- Creating new document sometimes didn't work (just in debugger?)
 
 ## [1.0.0] - 2021-09-21